+++ src/client/src/App.tsx
@@ -21,7 +21,7 @@ import { getVersionsDataAction } from "./store/versions/action";
 
 import appStyles from "./appStyles.module.css";
 import { IVersions } from "./types/version";
-import { IVSCodeObject } from "./store/vscode/model";
+import { IVSCodeObject } from "./types/vscode";
 import { IServiceStatus } from "./store/generationStatus/model";
 import { ISelected } from "./types/selected";
 import { AppState } from "./store/combineReducers";
