+++ src/client/src/containers/AppServiceModal/AppNameEditor/index.tsx
@@ -10,7 +10,7 @@ import { ValidateAppServiceName, GetValidAppServiceName } from "../../../utils/e
 import { useSelector } from "react-redux";
 import { AppState } from "../../../store/combineReducers";
 import { AppContext } from "../../../AppContext";
-import { IVSCodeObject } from "../../../store/vscode/model";
+import { IVSCodeObject } from "../../../types/vscode";
 
 let timeout: NodeJS.Timeout | undefined;
 let validationAppServiceNameScopeId = 0;
