+++ src/client/src/containers/RightSidebar/ServicesList/CosmosDBSelection/index.tsx
@@ -9,7 +9,7 @@ import styles from "./styles.module.css";
 import { KEY_EVENTS, EXTENSION_COMMANDS } from "../../../../utils/constants";
 import { injectIntl, InjectedIntlProps } from "react-intl";
 import { sendTelemetry } from "../../../../utils/extensionService/extensionService";
-import { IVSCodeObject } from "../../../../store/vscode/model";
+import { IVSCodeObject } from "../../../../types/vscode";
 import { removeCosmosSelectionAction } from "../../../../store/azureProfileData/cosmosDb/action";
 import { AppContext } from "../../../../AppContext";
