+++ src/client/src/types/vscode.ts
