+++ src/client/src/utils/validations/projectName/projectName.test.ts
@@ -14,7 +14,7 @@ jest.mock('../validations',()=>{return {
 }});
 import { validateProjectName } from "./projectName";
 import { IprojectNameValidationConfig } from "../../../reducers/wizardSelectionReducers/setValidationsAction";
-import { IVSCodeObject } from "../../../store/vscode/model";
+import { IVSCodeObject } from "../../../types/vscode";
 import { addRequiredValidate, addExistingProjectNameValidate, addRegexValidate,
   addReservedNameValidate } from '../validations';
